python-requirements.txt
# Keep sorted

jsonschema==4.23.0
mypy==1.14.1
numpy==2.2.2
pyyaml==6.0.2
scipy==1.15.1
tabulate==0.9.0
termcolor==2.5.0
